atc: structure: worker.Pool handles vol/container lookup
(and volume creation), rather than the worker.Client.

the eventual goal is to convert worker.Client into a client for a single
worker, rather than all the workers. the current flow for e.g. a `get`
step is:

exec.GetStep.Run(...)
    -> worker.Client.RunGetStep()
        -> worker.Pool.FindOrChooseWorkerForContainer()
            <- worker.Worker
        -> worker.Worker.Fetch(...)
            -> Garden specific stuff

the problem with this is that `worker.Worker` is Garden specific.
`worker.Client` is the ideal entrypoint to supporting multiple runtimes
(e.g. Ciro's K8s runtime PoC modified `worker.Client` to be k8s specific
instead), but since a `worker.Client` is the interface to interact with
ALL workers, that meant breaking Garden-based workers. if instead, the
flow went like:

exec.GetStep.Run(...)
    -> worker.Pool.FindOrChooseWorkerClientForContainer()
        <- worker.Client
    -> worker.Client.RunGetStep()
        -> worker.Worker.Fetch(...)
            -> Garden specific stuff

...then it's easy to swap out the `worker.Client` with a k8s (or nomad,
or whatever) implementation.

this also makes it possible to run multiple things on the same worker
explicitly. e.g. for my services RFC, it'd be nice to say:

exec.TaskStep.Run(...)
    -> worker.Pool.FindOrChooseWorkerClientForContainer()
        <- worker.Client
    -> worker.Client.RunService()
    -> worker.Client.RunTaskStep() // same worker
