+++ atc/api/containerserver/hijack.go
@@ -89,7 +89,7 @@ func (s *Server) HijackContainer(team db.Team) http.Handler {
 			"handle": handle,
 		})
 
-		container, found, err := s.workerClient.FindContainer(hLog, team.ID(), handle)
+		container, found, err := s.workerPool.FindContainer(hLog, team.ID(), handle)
 		if err != nil {
 			hLog.Error("failed-to-find-container", err)
 			w.WriteHeader(http.StatusInternalServerError)
