+++ atc/worker/client_test.go
@@ -22,7 +22,6 @@ import (
 	"github.com/onsi/gomega/gbytes"
 
 	"code.cloudfoundry.org/lager/lagertest"
-	"github.com/concourse/baggageclaim"
 	"github.com/concourse/concourse/atc/db"
 	"github.com/concourse/concourse/atc/worker"
 	"github.com/concourse/concourse/atc/worker/workerfakes"
@@ -35,7 +34,6 @@ var _ = Describe("Client", func() {
 	var (
 		logger          *lagertest.TestLogger
 		fakePool        *workerfakes.FakePool
-		fakeProvider    *workerfakes.FakeWorkerProvider
 		client          worker.Client
 		fakeLock        *lockfakes.FakeLock
 		fakeLockFactory *lockfakes.FakeLockFactory
@@ -45,183 +43,11 @@ var _ = Describe("Client", func() {
 	BeforeEach(func() {
 		logger = lagertest.NewTestLogger("test")
 		fakePool = new(workerfakes.FakePool)
-		fakeProvider = new(workerfakes.FakeWorkerProvider)
 		fakeCompression = new(compressionfakes.FakeCompression)
 		workerPolling := 1 * time.Second
 		workerStatus := 2 * time.Second
 
-		client = worker.NewClient(fakePool, fakeProvider, fakeCompression, workerPolling, workerStatus, false, 15*time.Minute)
-	})
-
-	Describe("FindContainer", func() {
-		var (
-			foundContainer worker.Container
-			found          bool
-			findErr        error
-		)
-
-		JustBeforeEach(func() {
-			foundContainer, found, findErr = client.FindContainer(
-				logger,
-				4567,
-				"some-handle",
-			)
-		})
-
-		Context("when looking up the worker errors", func() {
-			BeforeEach(func() {
-				fakeProvider.FindWorkerForContainerReturns(nil, false, errors.New("nope"))
-			})
-
-			It("errors", func() {
-				Expect(findErr).To(HaveOccurred())
-			})
-		})
-
-		Context("when worker is not found", func() {
-			BeforeEach(func() {
-				fakeProvider.FindWorkerForContainerReturns(nil, false, nil)
-			})
-
-			It("returns not found", func() {
-				Expect(findErr).NotTo(HaveOccurred())
-				Expect(found).To(BeFalse())
-			})
-		})
-
-		Context("when a worker is found with the container", func() {
-			var fakeWorker *workerfakes.FakeWorker
-			var fakeContainer *workerfakes.FakeContainer
-
-			BeforeEach(func() {
-				fakeWorker = new(workerfakes.FakeWorker)
-				fakeProvider.FindWorkerForContainerReturns(fakeWorker, true, nil)
-
-				fakeContainer = new(workerfakes.FakeContainer)
-				fakeWorker.FindContainerByHandleReturns(fakeContainer, true, nil)
-			})
-
-			It("succeeds", func() {
-				Expect(found).To(BeTrue())
-				Expect(findErr).NotTo(HaveOccurred())
-			})
-
-			It("returns the created container", func() {
-				Expect(foundContainer).To(Equal(fakeContainer))
-			})
-		})
-	})
-
-	Describe("FindVolume", func() {
-		var (
-			foundVolume worker.Volume
-			found       bool
-			findErr     error
-		)
-
-		JustBeforeEach(func() {
-			foundVolume, found, findErr = client.FindVolume(
-				logger,
-				4567,
-				"some-handle",
-			)
-		})
-
-		Context("when looking up the worker errors", func() {
-			BeforeEach(func() {
-				fakeProvider.FindWorkerForVolumeReturns(nil, false, errors.New("nope"))
-			})
-
-			It("errors", func() {
-				Expect(findErr).To(HaveOccurred())
-			})
-		})
-
-		Context("when worker is not found", func() {
-			BeforeEach(func() {
-				fakeProvider.FindWorkerForVolumeReturns(nil, false, nil)
-			})
-
-			It("returns not found", func() {
-				Expect(findErr).NotTo(HaveOccurred())
-				Expect(found).To(BeFalse())
-			})
-		})
-
-		Context("when a worker is found with the volume", func() {
-			var fakeWorker *workerfakes.FakeWorker
-			var fakeVolume *workerfakes.FakeVolume
-
-			BeforeEach(func() {
-				fakeWorker = new(workerfakes.FakeWorker)
-				fakeProvider.FindWorkerForVolumeReturns(fakeWorker, true, nil)
-
-				fakeVolume = new(workerfakes.FakeVolume)
-				fakeWorker.LookupVolumeReturns(fakeVolume, true, nil)
-			})
-
-			It("succeeds", func() {
-				Expect(found).To(BeTrue())
-				Expect(findErr).NotTo(HaveOccurred())
-			})
-
-			It("returns the volume", func() {
-				Expect(foundVolume).To(Equal(fakeVolume))
-			})
-		})
-	})
-
-	Describe("CreateVolume", func() {
-		var (
-			fakeWorker *workerfakes.FakeWorker
-			volumeSpec worker.VolumeSpec
-			workerSpec worker.WorkerSpec
-			volumeType db.VolumeType
-			err        error
-		)
-
-		BeforeEach(func() {
-			volumeSpec = worker.VolumeSpec{
-				Strategy: baggageclaim.EmptyStrategy{},
-			}
-
-			workerSpec = worker.WorkerSpec{
-				TeamID: 1,
-			}
-
-			volumeType = db.VolumeTypeArtifact
-		})
-
-		JustBeforeEach(func() {
-			_, err = client.CreateVolume(logger, volumeSpec, workerSpec, volumeType)
-		})
-
-		Context("when no workers can be found", func() {
-			BeforeEach(func() {
-				fakePool.FindOrChooseWorkerReturns(nil, errors.New("nope"))
-			})
-
-			It("returns an error", func() {
-				Expect(err).To(HaveOccurred())
-			})
-		})
-
-		Context("when the worker can be found", func() {
-			BeforeEach(func() {
-				fakeWorker = new(workerfakes.FakeWorker)
-				fakePool.FindOrChooseWorkerReturns(fakeWorker, nil)
-			})
-
-			It("creates the volume on the worker", func() {
-				Expect(err).ToNot(HaveOccurred())
-				Expect(fakeWorker.CreateVolumeCallCount()).To(Equal(1))
-				l, spec, id, t := fakeWorker.CreateVolumeArgsForCall(0)
-				Expect(l).To(Equal(logger))
-				Expect(spec).To(Equal(volumeSpec))
-				Expect(id).To(Equal(1))
-				Expect(t).To(Equal(volumeType))
-			})
-		})
+		client = worker.NewClient(fakePool, fakeCompression, workerPolling, workerStatus, false, 15*time.Minute)
 	})
 
 	Describe("RunCheckStep", func() {
@@ -301,7 +127,6 @@ var _ = Describe("Client", func() {
 
 			Describe("with an image artifact", func() {
 				var fakeArtifact *runtimefakes.FakeArtifact
-				var fakeVolumeWorker *workerfakes.FakeWorker
 				var fakeVolume *workerfakes.FakeVolume
 
 				BeforeEach(func() {
@@ -311,11 +136,8 @@ var _ = Describe("Client", func() {
 						ImageArtifact: fakeArtifact,
 					}
 
-					fakeVolumeWorker = new(workerfakes.FakeWorker)
-					fakeProvider.FindWorkerForVolumeReturns(fakeVolumeWorker, true, nil)
-
 					fakeVolume = new(workerfakes.FakeVolume)
-					fakeVolumeWorker.LookupVolumeReturns(fakeVolume, true, nil)
+					fakePool.FindVolumeReturns(fakeVolume, true, nil)
 				})
 
 				It("locates the volume and assigns it as an ImageArtifactSource", func() {
@@ -522,7 +344,6 @@ var _ = Describe("Client", func() {
 
 			Describe("with an image artifact", func() {
 				var fakeArtifact *runtimefakes.FakeArtifact
-				var fakeVolumeWorker *workerfakes.FakeWorker
 				var fakeVolume *workerfakes.FakeVolume
 
 				BeforeEach(func() {
@@ -532,11 +353,8 @@ var _ = Describe("Client", func() {
 						ImageArtifact: fakeArtifact,
 					}
 
-					fakeVolumeWorker = new(workerfakes.FakeWorker)
-					fakeProvider.FindWorkerForVolumeReturns(fakeVolumeWorker, true, nil)
-
 					fakeVolume = new(workerfakes.FakeVolume)
-					fakeVolumeWorker.LookupVolumeReturns(fakeVolume, true, nil)
+					fakePool.FindVolumeReturns(fakeVolume, true, nil)
 				})
 
 				It("locates the volume and assigns it as an ImageArtifactSource", func() {
@@ -809,7 +627,6 @@ var _ = Describe("Client", func() {
 
 		Describe("with an image artifact", func() {
 			var fakeArtifact *runtimefakes.FakeArtifact
-			var fakeVolumeWorker *workerfakes.FakeWorker
 			var fakeVolume *workerfakes.FakeVolume
 
 			BeforeEach(func() {
@@ -819,11 +636,8 @@ var _ = Describe("Client", func() {
 					ImageArtifact: fakeArtifact,
 				}
 
-				fakeVolumeWorker = new(workerfakes.FakeWorker)
-				fakeProvider.FindWorkerForVolumeReturns(fakeVolumeWorker, true, nil)
-
 				fakeVolume = new(workerfakes.FakeVolume)
-				fakeVolumeWorker.LookupVolumeReturns(fakeVolume, true, nil)
+				fakePool.FindVolumeReturns(fakeVolume, true, nil)
 			})
 
 			It("locates the volume and assigns it as an ImageArtifactSource", func() {
@@ -1494,7 +1308,6 @@ var _ = Describe("Client", func() {
 
 			Describe("with an image artifact", func() {
 				var fakeArtifact *runtimefakes.FakeArtifact
-				var fakeVolumeWorker *workerfakes.FakeWorker
 				var fakeVolume *workerfakes.FakeVolume
 
 				BeforeEach(func() {
@@ -1504,11 +1317,8 @@ var _ = Describe("Client", func() {
 						ImageArtifact: fakeArtifact,
 					}
 
-					fakeVolumeWorker = new(workerfakes.FakeWorker)
-					fakeProvider.FindWorkerForVolumeReturns(fakeVolumeWorker, true, nil)
-
 					fakeVolume = new(workerfakes.FakeVolume)
-					fakeVolumeWorker.LookupVolumeReturns(fakeVolume, true, nil)
+					fakePool.FindVolumeReturns(fakeVolume, true, nil)
 				})
 
 				It("locates the volume and assigns it as an ImageArtifactSource", func() {
