+++ atc/worker/pool_test.go
@@ -7,8 +7,11 @@ import (
 	"context"
 	"errors"
 
+	"github.com/concourse/baggageclaim"
 	"github.com/concourse/concourse/atc"
+	"github.com/concourse/concourse/atc/db"
 	"github.com/concourse/concourse/atc/db/dbfakes"
+	"github.com/concourse/concourse/atc/worker"
 	. "github.com/concourse/concourse/atc/worker"
 	"github.com/concourse/concourse/atc/worker/workerfakes"
 	. "github.com/onsi/ginkgo"
@@ -27,6 +30,178 @@ var _ = Describe("Pool", func() {
 		fakeProvider = new(workerfakes.FakeWorkerProvider)
 
 		pool = NewPool(fakeProvider)
+	})
+
+	Describe("FindContainer", func() {
+		var (
+			foundContainer worker.Container
+			found          bool
+			findErr        error
+		)
+
+		JustBeforeEach(func() {
+			foundContainer, found, findErr = pool.FindContainer(
+				logger,
+				4567,
+				"some-handle",
+			)
+		})
+
+		Context("when looking up the worker errors", func() {
+			BeforeEach(func() {
+				fakeProvider.FindWorkerForContainerReturns(nil, false, errors.New("nope"))
+			})
+
+			It("errors", func() {
+				Expect(findErr).To(HaveOccurred())
+			})
+		})
+
+		Context("when worker is not found", func() {
+			BeforeEach(func() {
+				fakeProvider.FindWorkerForContainerReturns(nil, false, nil)
+			})
+
+			It("returns not found", func() {
+				Expect(findErr).NotTo(HaveOccurred())
+				Expect(found).To(BeFalse())
+			})
+		})
+
+		Context("when a worker is found with the container", func() {
+			var fakeWorker *workerfakes.FakeWorker
+			var fakeContainer *workerfakes.FakeContainer
+
+			BeforeEach(func() {
+				fakeWorker = new(workerfakes.FakeWorker)
+				fakeProvider.FindWorkerForContainerReturns(fakeWorker, true, nil)
+
+				fakeContainer = new(workerfakes.FakeContainer)
+				fakeWorker.FindContainerByHandleReturns(fakeContainer, true, nil)
+			})
+
+			It("succeeds", func() {
+				Expect(found).To(BeTrue())
+				Expect(findErr).NotTo(HaveOccurred())
+			})
+
+			It("returns the created container", func() {
+				Expect(foundContainer).To(Equal(fakeContainer))
+			})
+		})
+	})
+
+	Describe("FindVolume", func() {
+		var (
+			foundVolume worker.Volume
+			found       bool
+			findErr     error
+		)
+
+		JustBeforeEach(func() {
+			foundVolume, found, findErr = pool.FindVolume(
+				logger,
+				4567,
+				"some-handle",
+			)
+		})
+
+		Context("when looking up the worker errors", func() {
+			BeforeEach(func() {
+				fakeProvider.FindWorkerForVolumeReturns(nil, false, errors.New("nope"))
+			})
+
+			It("errors", func() {
+				Expect(findErr).To(HaveOccurred())
+			})
+		})
+
+		Context("when worker is not found", func() {
+			BeforeEach(func() {
+				fakeProvider.FindWorkerForVolumeReturns(nil, false, nil)
+			})
+
+			It("returns not found", func() {
+				Expect(findErr).NotTo(HaveOccurred())
+				Expect(found).To(BeFalse())
+			})
+		})
+
+		Context("when a worker is found with the volume", func() {
+			var fakeWorker *workerfakes.FakeWorker
+			var fakeVolume *workerfakes.FakeVolume
+
+			BeforeEach(func() {
+				fakeWorker = new(workerfakes.FakeWorker)
+				fakeProvider.FindWorkerForVolumeReturns(fakeWorker, true, nil)
+
+				fakeVolume = new(workerfakes.FakeVolume)
+				fakeWorker.LookupVolumeReturns(fakeVolume, true, nil)
+			})
+
+			It("succeeds", func() {
+				Expect(found).To(BeTrue())
+				Expect(findErr).NotTo(HaveOccurred())
+			})
+
+			It("returns the volume", func() {
+				Expect(foundVolume).To(Equal(fakeVolume))
+			})
+		})
+	})
+
+	Describe("CreateVolume", func() {
+		var (
+			fakeWorker *workerfakes.FakeWorker
+			volumeSpec worker.VolumeSpec
+			workerSpec worker.WorkerSpec
+			volumeType db.VolumeType
+			err        error
+		)
+
+		BeforeEach(func() {
+			volumeSpec = worker.VolumeSpec{
+				Strategy: baggageclaim.EmptyStrategy{},
+			}
+
+			workerSpec = worker.WorkerSpec{
+				TeamID: 1,
+			}
+
+			volumeType = db.VolumeTypeArtifact
+		})
+
+		JustBeforeEach(func() {
+			_, err = pool.CreateVolume(logger, volumeSpec, workerSpec, volumeType)
+		})
+
+		Context("when no workers can be found", func() {
+			BeforeEach(func() {
+				fakeProvider.RunningWorkersReturns(nil, nil)
+			})
+
+			It("returns an error", func() {
+				Expect(err).To(HaveOccurred())
+			})
+		})
+
+		Context("when the worker can be found", func() {
+			BeforeEach(func() {
+				fakeWorker = new(workerfakes.FakeWorker)
+				fakeProvider.RunningWorkersReturns([]Worker{fakeWorker}, nil)
+				fakeWorker.SatisfiesReturns(true)
+			})
+
+			It("creates the volume on the worker", func() {
+				Expect(err).ToNot(HaveOccurred())
+				Expect(fakeWorker.CreateVolumeCallCount()).To(Equal(1))
+				l, spec, id, t := fakeWorker.CreateVolumeArgsForCall(0)
+				Expect(l).To(Equal(logger))
+				Expect(spec).To(Equal(volumeSpec))
+				Expect(id).To(Equal(1))
+				Expect(t).To(Equal(volumeType))
+			})
+		})
 	})
 
 	Describe("FindOrChooseWorkerForContainer", func() {
