+++ atc/worker/workerfakes/fake_client.go
@@ -15,56 +15,6 @@ import (
 )
 
 type FakeClient struct {
-	CreateVolumeStub        func(lager.Logger, worker.VolumeSpec, worker.WorkerSpec, db.VolumeType) (worker.Volume, error)
-	createVolumeMutex       sync.RWMutex
-	createVolumeArgsForCall []struct {
-		arg1 lager.Logger
-		arg2 worker.VolumeSpec
-		arg3 worker.WorkerSpec
-		arg4 db.VolumeType
-	}
-	createVolumeReturns struct {
-		result1 worker.Volume
-		result2 error
-	}
-	createVolumeReturnsOnCall map[int]struct {
-		result1 worker.Volume
-		result2 error
-	}
-	FindContainerStub        func(lager.Logger, int, string) (worker.Container, bool, error)
-	findContainerMutex       sync.RWMutex
-	findContainerArgsForCall []struct {
-		arg1 lager.Logger
-		arg2 int
-		arg3 string
-	}
-	findContainerReturns struct {
-		result1 worker.Container
-		result2 bool
-		result3 error
-	}
-	findContainerReturnsOnCall map[int]struct {
-		result1 worker.Container
-		result2 bool
-		result3 error
-	}
-	FindVolumeStub        func(lager.Logger, int, string) (worker.Volume, bool, error)
-	findVolumeMutex       sync.RWMutex
-	findVolumeArgsForCall []struct {
-		arg1 lager.Logger
-		arg2 int
-		arg3 string
-	}
-	findVolumeReturns struct {
-		result1 worker.Volume
-		result2 bool
-		result3 error
-	}
-	findVolumeReturnsOnCall map[int]struct {
-		result1 worker.Volume
-		result2 bool
-		result3 error
-	}
 	RunCheckStepStub        func(context.Context, lager.Logger, db.ContainerOwner, worker.ContainerSpec, worker.WorkerSpec, worker.ContainerPlacementStrategy, db.ContainerMetadata, runtime.ProcessSpec, runtime.StartingEventDelegate, resource.Resource) (worker.CheckResult, error)
 	runCheckStepMutex       sync.RWMutex
 	runCheckStepArgsForCall []struct {
@@ -172,208 +122,6 @@ type FakeClient struct {
 	}
 	invocations      map[string][][]interface{}
 	invocationsMutex sync.RWMutex
-}
-
-func (fake *FakeClient) CreateVolume(arg1 lager.Logger, arg2 worker.VolumeSpec, arg3 worker.WorkerSpec, arg4 db.VolumeType) (worker.Volume, error) {
-	fake.createVolumeMutex.Lock()
-	ret, specificReturn := fake.createVolumeReturnsOnCall[len(fake.createVolumeArgsForCall)]
-	fake.createVolumeArgsForCall = append(fake.createVolumeArgsForCall, struct {
-		arg1 lager.Logger
-		arg2 worker.VolumeSpec
-		arg3 worker.WorkerSpec
-		arg4 db.VolumeType
-	}{arg1, arg2, arg3, arg4})
-	fake.recordInvocation("CreateVolume", []interface{}{arg1, arg2, arg3, arg4})
-	fake.createVolumeMutex.Unlock()
-	if fake.CreateVolumeStub != nil {
-		return fake.CreateVolumeStub(arg1, arg2, arg3, arg4)
-	}
-	if specificReturn {
-		return ret.result1, ret.result2
-	}
-	fakeReturns := fake.createVolumeReturns
-	return fakeReturns.result1, fakeReturns.result2
-}
-
-func (fake *FakeClient) CreateVolumeCallCount() int {
-	fake.createVolumeMutex.RLock()
-	defer fake.createVolumeMutex.RUnlock()
-	return len(fake.createVolumeArgsForCall)
-}
-
-func (fake *FakeClient) CreateVolumeCalls(stub func(lager.Logger, worker.VolumeSpec, worker.WorkerSpec, db.VolumeType) (worker.Volume, error)) {
-	fake.createVolumeMutex.Lock()
-	defer fake.createVolumeMutex.Unlock()
-	fake.CreateVolumeStub = stub
-}
-
-func (fake *FakeClient) CreateVolumeArgsForCall(i int) (lager.Logger, worker.VolumeSpec, worker.WorkerSpec, db.VolumeType) {
-	fake.createVolumeMutex.RLock()
-	defer fake.createVolumeMutex.RUnlock()
-	argsForCall := fake.createVolumeArgsForCall[i]
-	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3, argsForCall.arg4
-}
-
-func (fake *FakeClient) CreateVolumeReturns(result1 worker.Volume, result2 error) {
-	fake.createVolumeMutex.Lock()
-	defer fake.createVolumeMutex.Unlock()
-	fake.CreateVolumeStub = nil
-	fake.createVolumeReturns = struct {
-		result1 worker.Volume
-		result2 error
-	}{result1, result2}
-}
-
-func (fake *FakeClient) CreateVolumeReturnsOnCall(i int, result1 worker.Volume, result2 error) {
-	fake.createVolumeMutex.Lock()
-	defer fake.createVolumeMutex.Unlock()
-	fake.CreateVolumeStub = nil
-	if fake.createVolumeReturnsOnCall == nil {
-		fake.createVolumeReturnsOnCall = make(map[int]struct {
-			result1 worker.Volume
-			result2 error
-		})
-	}
-	fake.createVolumeReturnsOnCall[i] = struct {
-		result1 worker.Volume
-		result2 error
-	}{result1, result2}
-}
-
-func (fake *FakeClient) FindContainer(arg1 lager.Logger, arg2 int, arg3 string) (worker.Container, bool, error) {
-	fake.findContainerMutex.Lock()
-	ret, specificReturn := fake.findContainerReturnsOnCall[len(fake.findContainerArgsForCall)]
-	fake.findContainerArgsForCall = append(fake.findContainerArgsForCall, struct {
-		arg1 lager.Logger
-		arg2 int
-		arg3 string
-	}{arg1, arg2, arg3})
-	fake.recordInvocation("FindContainer", []interface{}{arg1, arg2, arg3})
-	fake.findContainerMutex.Unlock()
-	if fake.FindContainerStub != nil {
-		return fake.FindContainerStub(arg1, arg2, arg3)
-	}
-	if specificReturn {
-		return ret.result1, ret.result2, ret.result3
-	}
-	fakeReturns := fake.findContainerReturns
-	return fakeReturns.result1, fakeReturns.result2, fakeReturns.result3
-}
-
-func (fake *FakeClient) FindContainerCallCount() int {
-	fake.findContainerMutex.RLock()
-	defer fake.findContainerMutex.RUnlock()
-	return len(fake.findContainerArgsForCall)
-}
-
-func (fake *FakeClient) FindContainerCalls(stub func(lager.Logger, int, string) (worker.Container, bool, error)) {
-	fake.findContainerMutex.Lock()
-	defer fake.findContainerMutex.Unlock()
-	fake.FindContainerStub = stub
-}
-
-func (fake *FakeClient) FindContainerArgsForCall(i int) (lager.Logger, int, string) {
-	fake.findContainerMutex.RLock()
-	defer fake.findContainerMutex.RUnlock()
-	argsForCall := fake.findContainerArgsForCall[i]
-	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3
-}
-
-func (fake *FakeClient) FindContainerReturns(result1 worker.Container, result2 bool, result3 error) {
-	fake.findContainerMutex.Lock()
-	defer fake.findContainerMutex.Unlock()
-	fake.FindContainerStub = nil
-	fake.findContainerReturns = struct {
-		result1 worker.Container
-		result2 bool
-		result3 error
-	}{result1, result2, result3}
-}
-
-func (fake *FakeClient) FindContainerReturnsOnCall(i int, result1 worker.Container, result2 bool, result3 error) {
-	fake.findContainerMutex.Lock()
-	defer fake.findContainerMutex.Unlock()
-	fake.FindContainerStub = nil
-	if fake.findContainerReturnsOnCall == nil {
-		fake.findContainerReturnsOnCall = make(map[int]struct {
-			result1 worker.Container
-			result2 bool
-			result3 error
-		})
-	}
-	fake.findContainerReturnsOnCall[i] = struct {
-		result1 worker.Container
-		result2 bool
-		result3 error
-	}{result1, result2, result3}
-}
-
-func (fake *FakeClient) FindVolume(arg1 lager.Logger, arg2 int, arg3 string) (worker.Volume, bool, error) {
-	fake.findVolumeMutex.Lock()
-	ret, specificReturn := fake.findVolumeReturnsOnCall[len(fake.findVolumeArgsForCall)]
-	fake.findVolumeArgsForCall = append(fake.findVolumeArgsForCall, struct {
-		arg1 lager.Logger
-		arg2 int
-		arg3 string
-	}{arg1, arg2, arg3})
-	fake.recordInvocation("FindVolume", []interface{}{arg1, arg2, arg3})
-	fake.findVolumeMutex.Unlock()
-	if fake.FindVolumeStub != nil {
-		return fake.FindVolumeStub(arg1, arg2, arg3)
-	}
-	if specificReturn {
-		return ret.result1, ret.result2, ret.result3
-	}
-	fakeReturns := fake.findVolumeReturns
-	return fakeReturns.result1, fakeReturns.result2, fakeReturns.result3
-}
-
-func (fake *FakeClient) FindVolumeCallCount() int {
-	fake.findVolumeMutex.RLock()
-	defer fake.findVolumeMutex.RUnlock()
-	return len(fake.findVolumeArgsForCall)
-}
-
-func (fake *FakeClient) FindVolumeCalls(stub func(lager.Logger, int, string) (worker.Volume, bool, error)) {
-	fake.findVolumeMutex.Lock()
-	defer fake.findVolumeMutex.Unlock()
-	fake.FindVolumeStub = stub
-}
-
-func (fake *FakeClient) FindVolumeArgsForCall(i int) (lager.Logger, int, string) {
-	fake.findVolumeMutex.RLock()
-	defer fake.findVolumeMutex.RUnlock()
-	argsForCall := fake.findVolumeArgsForCall[i]
-	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3
-}
-
-func (fake *FakeClient) FindVolumeReturns(result1 worker.Volume, result2 bool, result3 error) {
-	fake.findVolumeMutex.Lock()
-	defer fake.findVolumeMutex.Unlock()
-	fake.FindVolumeStub = nil
-	fake.findVolumeReturns = struct {
-		result1 worker.Volume
-		result2 bool
-		result3 error
-	}{result1, result2, result3}
-}
-
-func (fake *FakeClient) FindVolumeReturnsOnCall(i int, result1 worker.Volume, result2 bool, result3 error) {
-	fake.findVolumeMutex.Lock()
-	defer fake.findVolumeMutex.Unlock()
-	fake.FindVolumeStub = nil
-	if fake.findVolumeReturnsOnCall == nil {
-		fake.findVolumeReturnsOnCall = make(map[int]struct {
-			result1 worker.Volume
-			result2 bool
-			result3 error
-		})
-	}
-	fake.findVolumeReturnsOnCall[i] = struct {
-		result1 worker.Volume
-		result2 bool
-		result3 error
-	}{result1, result2, result3}
 }
 
 func (fake *FakeClient) RunCheckStep(arg1 context.Context, arg2 lager.Logger, arg3 db.ContainerOwner, arg4 worker.ContainerSpec, arg5 worker.WorkerSpec, arg6 worker.ContainerPlacementStrategy, arg7 db.ContainerMetadata, arg8 runtime.ProcessSpec, arg9 runtime.StartingEventDelegate, arg10 resource.Resource) (worker.CheckResult, error) {
@@ -734,12 +482,6 @@ func (fake *FakeClient) StreamFileFromArtifactReturnsOnCall(i int, result1 io.Re
 func (fake *FakeClient) Invocations() map[string][][]interface{} {
 	fake.invocationsMutex.RLock()
 	defer fake.invocationsMutex.RUnlock()
-	fake.createVolumeMutex.RLock()
-	defer fake.createVolumeMutex.RUnlock()
-	fake.findContainerMutex.RLock()
-	defer fake.findContainerMutex.RUnlock()
-	fake.findVolumeMutex.RLock()
-	defer fake.findVolumeMutex.RUnlock()
 	fake.runCheckStepMutex.RLock()
 	defer fake.runCheckStepMutex.RUnlock()
 	fake.runGetStepMutex.RLock()
